+++ packages/badge-api/badge/test/badge/handler.spec.ts
@@ -4,11 +4,11 @@ import { expect } from 'chai';
 import type { SinonStubbedInstance } from 'sinon';
 import sinon from 'sinon';
 
-import { handler } from '../../badge/handler.js';
-import type { Shield } from '../../badge/Shield.js';
-import { Color } from '../../badge/Shield.js';
-import { ShieldMapper } from '../../badge/ShieldMapper.js';
-import pkg from '../../package.json' with { type: 'json' };
+import pkg from '../../../package.json' with { type: 'json' };
+import { handler } from '../../handler.js';
+import type { Shield } from '../../Shield.js';
+import { Color } from '../../Shield.js';
+import { ShieldMapper } from '../../ShieldMapper.js';
 
 const headers = {
   ['X-Badge-Api-Version']: pkg.version,
