+++ packages/website-backend/bin/dashboard-backend.js
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
-import '../dist/src/index.js';
+import '../dist/index.js';
 
 process.title = 'dashboard';
