+++ packages/website-backend/src/test/unit/client/HttpClientSpec.ts
@@ -1,8 +1,8 @@
 import { expect } from 'chai';
 import sinon from 'sinon';
 
-import type { Response } from '../../../src/client/HttpClient.js';
-import HttpClient from '../../../src/client/HttpClient.js';
+import type { Response } from '../../../client/HttpClient.js';
+import HttpClient from '../../../client/HttpClient.js';
 
 describe('HttpClient', () => {
   let sut: HttpClient;
