+++ packages/website-backend/src/test/unit/controllers/repositories.controller.spec.ts
@@ -4,12 +4,12 @@ import sinon from 'sinon';
 import request from 'supertest';
 import type { App } from 'supertest/types.js';
 
-import { AppModule } from '../../../src/app.module.js';
-import type * as github from '../../../src/github/models.js';
-import Configuration from '../../../src/services/Configuration.js';
-import DataAccess from '../../../src/services/DataAccess.js';
-import GithubRepositoryService from '../../../src/services/GithubRepositoryService.js';
-import utils from '../../../src/utils/utils.js';
+import { AppModule } from '../../../app.module.js';
+import type * as github from '../../../github/models.js';
+import Configuration from '../../../services/Configuration.js';
+import DataAccess from '../../../services/DataAccess.js';
+import GithubRepositoryService from '../../../services/GithubRepositoryService.js';
+import utils from '../../../utils/utils.js';
 import { githubFactory } from '../../helpers/producers.js';
 import { config, createAuthorizationHeader, DataAccessMock } from '../../helpers/TestServer.js';
