+++ packages/website-backend/src/test/unit/controllers/user.controller.spec.ts
@@ -5,13 +5,13 @@ import sinon from 'sinon';
 import request from 'supertest';
 import type { App } from 'supertest/types.js';
 
-import { AppModule } from '../../../src/app.module.js';
-import UserController from '../../../src/controllers/user.controller.js';
-import GithubAgent from '../../../src/github/GithubAgent.js';
-import type { Authentication } from '../../../src/github/models.js';
-import Configuration from '../../../src/services/Configuration.js';
-import DataAccess from '../../../src/services/DataAccess.js';
-import GithubRepositoryService from '../../../src/services/GithubRepositoryService.js';
+import { AppModule } from '../../../app.module.js';
+import UserController from '../../../controllers/user.controller.js';
+import GithubAgent from '../../../github/GithubAgent.js';
+import type { Authentication } from '../../../github/models.js';
+import Configuration from '../../../services/Configuration.js';
+import DataAccess from '../../../services/DataAccess.js';
+import GithubRepositoryService from '../../../services/GithubRepositoryService.js';
 import { contractFactory, githubFactory } from '../../helpers/producers.js';
 import { config, createAuthorizationHeader, DataAccessMock } from '../../helpers/TestServer.js';
