+++ packages/website-backend/src/test/unit/github/GithubAgentSpec.ts
@@ -1,10 +1,10 @@
 import { expect } from 'chai';
 import sinon from 'sinon';
 
-import type { Response } from '../../../src/client/HttpClient.js';
-import HttpClient from '../../../src/client/HttpClient.js';
-import GithubAgent from '../../../src/github/GithubAgent.js';
-import type * as github from '../../../src/github/models.js';
+import type { Response } from '../../../client/HttpClient.js';
+import HttpClient from '../../../client/HttpClient.js';
+import GithubAgent from '../../../github/GithubAgent.js';
+import type * as github from '../../../github/models.js';
 import { githubFactory } from '../../helpers/producers.js';
 
 describe('GithubClient', () => {
