+++ packages/website-backend/src/test/unit/services/Configuration.spec.ts
@@ -1,8 +1,8 @@
 import { expect } from 'chai';
 import sinon from 'sinon';
 
-import Configuration from '../../../src/services/Configuration.js';
-import utils from '../../../src/utils/utils.js';
+import Configuration from '../../../services/Configuration.js';
+import utils from '../../../utils/utils.js';
 
 describe('Configuration', () => {
   let requiredEnvVarStub: sinon.SinonStub<[string], string>;
