+++ packages/website-backend/src/test/unit/services/real-time/MutationEventSender.spec.ts
@@ -2,7 +2,7 @@ import type { Response } from 'express';
 import type { MutantResult } from 'mutation-testing-report-schema';
 import sinon from 'sinon';
 
-import { MutationEventSender } from '../../../../src/services/real-time/MutationEventSender.js';
+import { MutationEventSender } from '../../../../services/real-time/MutationEventSender.js';
 import { createResponseStub } from '../../helpers.js';
 
 describe(MutationEventSender.name, () => {
