+++ packages/website-backend/stryker.conf.js
@@ -1,6 +1,5 @@
-import fs from 'fs';
+import settings from '../../stryker.parent.conf.json' with { type: 'json' };
 
-const settings = JSON.parse(fs.readFileSync(new URL('../../stryker.parent.conf.json', import.meta.url), 'utf-8'));
 settings.dashboard.module = import.meta.url.split('/').slice(-2)[0];
 
 /**
