+++ packages/website-backend/tsconfig.json
@@ -5,7 +5,7 @@
       "path": "src"
     },
     {
-      "path": "test"
+      "path": "src/test"
     }
   ]
 }
